python-services/scraper-service.py
#!/usr/bin/env python3
"""
Simple Python Scraper Microservice
This is a stub implementation that demonstrates how the Go backend
would communicate with the Python scraper service.
"""

from flask import Flask, request, jsonify
import logging
import time
import random

app = Flask(__name__)
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

@app.route('/scrape', methods=['POST'])
def scrape_seller():
    """
    Endpoint to trigger scraping for a seller.
    In the real implementation, this would call the existing Django scraper code.
    """
    try:
        data = request.get_json()
        seller_name = data.get('seller_name')
        
        if not seller_name:
            return jsonify({
                'success': False,
                'error': 'seller_name is required'
            }), 400
        
        logger.info(f"Starting scrape for seller: {seller_name}")
        
        # Simulate scraping work
        time.sleep(1)  # Simulate processing time
        
        # Simulate success/failure
        if random.random() > 0.1:  # 90% success rate
            logger.info(f"Scrape completed successfully for {seller_name}")
            return jsonify({
                'success': True,
                'message': f'Successfully scraped data for seller {seller_name}',
                'records_processed': random.randint(10, 100)
            })
        else:
            logger.error(f"Scrape failed for {seller_name}")
            return jsonify({
                'success': False,
                'error': f'Failed to scrape data for seller {seller_name}'
            }), 500
            
    except Exception as e:
        logger.error(f"Error in scrape endpoint: {str(e)}")
        return jsonify({
            'success': False,
            'error': f'Internal server error: {str(e)}'
        }), 500

@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'service': 'scraper',
        'timestamp': time.time()
    })

if __name__ == '__main__':
    logger.info("Starting Scraper Microservice on port 8001")
    app.run(host='0.0.0.0', port=8001, debug=True)
